Create SQL file
This query looks at all the parts we have demand for, but no supply. The ERP system will autogenerate supply suggestions only if the part has an approved revision. 

Approved parts with demand have jobs. The ERP system generates jobs to meet demand. I looked at all the parts and filtered based on those that are unapproved, have no jobs, and have open demand (open sales order release for the part number). 

The information provided includes when the next shipment is for the part and what the material requirement is for this part.